+++ native/cocos/renderer/gfx-metal/MTLDescriptorSet.mm
@@ -71,11 +71,11 @@ void CCMTLDescriptorSet::update() {
     if (_isDirty && _gpuDescriptorSet) {
         const auto &descriptors = _gpuDescriptorSet->gpuDescriptors;
         for (size_t i = 0; i < descriptors.size(); i++) {
-            if (static_cast<uint>(descriptors[i].type) & DESCRIPTOR_BUFFER_TYPE) {
+            if (descriptors[i].type & DESCRIPTOR_BUFFER_TYPE) {
                 if (_buffers[i]) {
                     _gpuDescriptorSet->gpuDescriptors[i].buffer = static_cast<CCMTLBuffer *>(_buffers[i]);
                 }
-            } else if (static_cast<uint>(descriptors[i].type) & DESCRIPTOR_SAMPLER_TYPE) {
+            } else if (descriptors[i].type & DESCRIPTOR_SAMPLER_TYPE) {
                 if (_textures[i]) {
                     _gpuDescriptorSet->gpuDescriptors[i].texture = static_cast<CCMTLTexture *>(_textures[i]);
                 }
